Put client container IP into enode:// URL
When the simulation asks for the client's enode URL, the hive controller
runs enode.sh in the client container and returns the output of that in
the HTTP response.

But if the client doesn't know what its own IP is, the returned URL
won't actually work. To make it work, parse the URL in hive and modify
it to contain the container IP.