clients/nethermind: simplify chain configuration
The chain configuration was created in a very weird way: the startup
script first created a geth-like chain config object, then wrote that
to the "config" key in the genesis.json file. Finally, the result was
filtered through mapper.jq, which ignored all this config data and
applied the forks from the environment instead. The jq script also
added in some defaults from the test.json config file.

This removes all the extra steps and just generates the genesis and
chain config in mapper.jq.

In addition, this also fixes the bootnode setup and disables the
external IP lookup.