simulators/ethereum/eth: add eth protocol test suite (#324)
This change adds a new simulator for the 'eth' peer-to-peer protocol
test suite. At this time, the test passes with geth and besu but fails with
all other clients because they have issues loading the test chain.

This change also simplifies the handling of enode URLs. When the simulation
asks for the client's enode URL, the hive controller runs enode.sh in the client
container and returns the output of that in the HTTP response.

But if the client doesn't know what its own IP is, the returned URL
won't actually work. To make it work, we now parse the URL in hive and
modify it to contain the container IP.