cmd/hiveview: fix navigation in the app
This was my number one issue with it: when loading a test suite, the
navigation would be stuck on it forever until clicking another suite.

If I wanted to check out a test suite, then go back to the main page and
reload to get new results, it didn't work, because reloading the app
would take me right back to the suite I previously clicked on.

This also fixes the forward/back buttons.