cmd/hiveview: fix more navigation issues (#351)
* cmd/hiveview: fix opening test suite more than once

This makes it possible to click 'Load' when a test suite was already
loaded once. This issue was always present, but we didn't notice because
it was impossible to navigate back and forth before the navigation fix.

* cmd/hiveview: show spinner in fileviewer

This reworks the file viewer to display a loading indicator. It can
actually take a long time to fetch big log files. Browser history
interaction is also fixed by this, and the page now sets the document
title to the file name after loading the file.

* cmd/hiveview: open home tab if page is not set

This fixes one more case where the selected tab could get out of sync
with the expected browser history behavior.