cmd/hivechain: add support for generating non-empty chain (#352)
The tool already had a function for generating non-empty chains, but it
didn't work with user-supplied genesis blocks. This change makes it work
by checking whether the genesis allocation assigns balance to certain
known addresses. If these addresses have ether, we can spend it in the
generator.

Several kinds of transactions are generated into the test chain:

- simple value transfers to random destination
- contract creations with some storage
- contract creations with random logs
- contract creations that deploy random code