simulators/smoke/network: use package hivesim
This makes the networking smoke test use the new hivesim package, which
simplifies the code a lot. For now, the simulation module needs to be
called 'network2' in order to avoid a compilation error about to
'ambiguous imports'. This error happens because the simulation's package
main was previously part of the hive module. I'll remove the '2' from
the name later.