simulators/dev2p/eth: integrate eth protocol TAP output
This changes the eth simulation to use the new TAP output feature of the
test tool. The simulation now reports results from individual protocol
tests to hive.

The simulation now uses package hivesim and is now a stand-alone Go
module. I also moved it to simulators/devp2p because we wanted to have
it there anyway, and renaming it now makes the Go module migration
easier.