simulators/devp2p/eth: integrate eth protocol test TAP output (#378)
This changes the eth simulation to use the new TAP output feature of the
test tool. The simulation now reports results from individual protocol
tests to hive.

The simulation now uses package hivesim and is a stand-alone Go
module. I also moved it to simulators/devp2p because we wanted to have
it there anyway, and renaming it in this PR makes the Go module migration
easier.