cmd/hiveview: show test details as preformatted text, not markdown (#379)
The markdown is nice, but requires a lot of effort in the test to format
it correctly. With the newer simulation API client where you can basically
write simulations like Go unit tests, outputting lines with t.Log(...) gets really
messy if it has to generate valid markdown.

This simplifies it a bit by just showing the text with <pre>.

I think the most useful feature of the markdown conversion was being able
to link to stuff. This is preserved with this change by making all URLs in the
output clickable.