simulators/ethereum/consensus: use package hivesim
This makes the consensus tests use the new 'hivesim' simulation API
client. The simulation is now a stand-alone module and no longer
requires the --sim.rootcontext option. While here, I also fixed some
staticcheck warnings in the simulation code.

Note that the simulation will behave slightly differently from before
when run with multiple client types: before this change, a separate
suite was started for every client. With the new version, the simulation
will only create a single suite and put all client test results into it.

This behavior change will not be noticed on the production Hive instance
though, be cause we invoke the simulation separately for every client.

I've tested this using the following command, and it does work as expected:

  ./hive --sim consensus --client go-ethereum --docker-noshell --sim.parallelism 8 --sim.testlimit 50