hivesim: add client start options (#442)
This breaking change to the test API adds 'start options'. When starting
a client instance, environment variables and files could be specified, but
the API for passing files to the client required storing the files in the
simulator's file system first. This is inconvenient when there are many files.
With start options, the client files can be created in memory instead.

The new API also makes it possible to combine multiple start options into
a single option value. This makes it easier to write test helper functions
which create a client configuration to be applied later.

Co-authored-by: Felix Lange <fjl@twurst.com>