internal/libhive: remove special handling for enode.sh
Since all simulators now request the enode URL using a /hive-bin script,
the API endpoint for running /enode.sh can be removed.

The API endpoint for method GET of a client is now unused, but may
become useful later. I have replaced it with a generic handler that
returns the client ID and name.