hivesim, internal/libhive: change 'start client' request input to JSON
This is another API change. As of this change, client launch parameters
are no longer submitted as HTML form parameters. Instead, a single form
parameter is used to carry a JSON object with the config.

I'm making this change mostly because it will be easier to extend the
interface when it is JSON-based.