simulators/ethereum/rpc: bump wait count in createAndFundAccount
When running tests against an LES client, the client may sometimes
connect and then immediately disconnect from the server because it
considers the server to be not ready.

When this happens, the client takes up to 30s to reconnect, and this
exceeds the waiting time in createAndFundAccount.

Fix it by making the test wait longer for inclusion.