hive: serve simulation API via proxy container (#545)
In order to communicate with the simulator container, hive needs to provide
an API listener on the same network as the simulator. The way this was done
is error-prone. At startup, it tried to figure out the local network
interface corresponding to the Docker 'bridge' network and listened there.
This worked on Linux, sometimes. Quite a few people have run into issues
with it.

This change makes API communication more portable. hive now listens to the
simulator via a 'proxy' container, which is an HTTP server that forwards
requests to hive via its stdout/stdin streams. Hive also delegates
container liveness checks to the proxy, because those checks also require
connectivity on the Docker network.