hive: move simulation runner to internal/libhive
This moves all logic related to a simulation run into the library. The
runner has also gained a couple of new features in this change:
PAss/fail stats are now collected across simulators, enabling the hive
tool to exit with a non-zero status when tests have failed.

The client list available to the simulator can now be overridden for
every simulation run.