hive: move simulation runner to internal/libhive (#601)
This moves all logic related to a simulation run into the library. The
runner has also gained a couple of new features in this change:
Pass/fail stats are now collected across the entire run, enabling
the hive tool to exit with a non-zero status when any tests have failed.

Another new feature of the library: the client list can now be overridden
separately for every individual simulator run. This will come in handy when
we create a tool to run multiple simulations from a config file.