Attempt to fix deposit/withdrawal tests (#48)
I narrowed this issue down to reverts during deposits. It seems like the default gas limit calculation under-estimates the limit sometimes which leads to deposits reverting without a reason. I'm not 100% sure why this is, but I have a hunch it's related to the `ResourceMetering` contract since specifying a high manual gas limit resolved the issue. I'll follow up with @tynes to dig deeper into whether this is correct, and understand the root cause.

To test this, I ran the tests 3 times against CI and got passes each time.

Fixes ENG-2726