all: Go workspace support and dependency sync (#664)
This change adds a top-level go.work file containing all Go-based simulators
as well as the hive module. Joining the modules in a workspace makes it
possible to easily share code among simulators.

When simulators share code, the imported simulator modules need to be
made available in the docker build environment. To facilitate this, the docker
image build context can now be configured using a special file in the simulator
directory.

Co-authored-by: Felix Lange <fjl@twurst.com>