clients/besu: fix clique mining (#730)
Reversed the logic to deploy clique key to bootnode only.
Original change was to avoid deploying the same clique key to 2 different
nodes. Ended up stopping the deployment of the clique key to the first node,
causing test that rely on a single node to mine to fail due to timeout.