libhive: client build args and alternative dockerfiles (#767)
This adds support for reading the client list from a YAML file. When using the clients file,
an alternative Dockerfile can be selected to build the client from source, for example.
It is also possible to set arbitrary build arguments for each client in the YAML file.

The "branch" build argument has been renamed to "tag". The image building code will
still set "branch" for temporary backwards-compatibility, but I plan to remove that later.

---------

Co-authored-by: Felix Lange <fjl@twurst.com>