cmd/hiveview, internal/libhive: lazy loading of test details (#800)
This changes the hive runner to store test logs in a separate file. Recent test suites for
EIP-4844 create > 200MB of log output for a single run, so it's no longer viable to store
all output in a single JSON file. So this PR makes hive store it in a dedicated test log
file instead.

The hiveview frontend is also updated to lazy-load test logs. Loading happens when the
test is opened. Some numbers: for the engine-withdrawals suite, the JSON output file was
~65MB in size. With this PR applied, the suite file is just 25KB, i.e. it loads instantly.