simulators/portal/beacon Remove incomplete gossip/offer tests
In order to test offer/gossip in the Portal beacon network,
the node needs to first also do light client sync to be able to
validate the offered content. This requires these test to be
adjusted to provide the proper light client data first in order
to be able to sync. Remove these incomplete tests for now to
avoid false failures.