refactor(nethermind): Refactor enode parsing to not use logs
Using logs is too fragile. It's not ideal to get the enode
from the logs if loglevel is set lower. All other clients
use the admin_nodeInfo for this purpose and we should do so
for Nethermind as well.

This also turns on admin module as this was necessary for this
change.